fix: separation of create mfa and list factors
